+++ src/main/frontend/app/views/about/about.component.ts
@@ -1,11 +1,5 @@
 import {Component, OnInit} from 'angular2/core';
 
-/*
- * We're loading this component asynchronously
- * We are using some magic with es6-promise-loader that will wrap the module with a Promise
- * see https://github.com/gdi2290/es6-promise-loader for more info
- */
-
 console.log('`About` component loaded asynchronously');
 
 @Component({